backend/llms/gemini_handler.py
import os
from dotenv import load_dotenv

class GeminiHandler:
    def __init__(self):
        # Load environment variables
        load_dotenv('.env.local')
        api_key = os.getenv('GEMINI_API_KEY')
        if not api_key:
            raise ValueError("GEMINI_API_KEY not found in environment variables")
        # Gemini model will be initialized here later
        
    def identify_car(self, image):
        """
        Identify a car using Google's Gemini Pro Vision model
        This is a placeholder that will be implemented later
        """
        return "Gemini identification not implemented yet" 